+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.5.0 (unreleased)
 ------------------
 
+- #2279 Allow all custom transitions in sample report listing
 - #2280 Remove custom date rendering in sample header
 - #2278 Client catalog
 - #2276 Senaite labels